requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
plotly>=5.15.0

# Graph neural networks
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1

# Evaluation and metrics
scikit-learn>=1.3.0
torchmetrics>=1.0.0

# Visualization
networkx>=3.1
pyvis>=0.3.2

# Optional logging
wandb>=0.15.0
tensorboard>=2.13.0

# Development
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0

# Demo
streamlit>=1.25.0
gradio>=3.40.0

# Optional serving
fastapi>=0.100.0
uvicorn>=0.23.0
